workflow_manager/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
httpx[http2]